==> p6-hotel-feature-labeling-and-rating-prediction/jupyter-notebook/Part01_in_capstone_convert_date_type.ipynb <==
## Outline of steps
* [step0](#step0): import necessary packages
* [step1](#step1): import dataset as `part1_dataset`
* [step2](#step2): have a look at the dataset and its data type
* [step3](#step3): convert `Review_Date` from object to datetime
* [step4](#step4): have a look at the descriptive statistical summary
* [step5](#step5): save the dataset as part1_dataset

# import necessary packages
import numpy as np
import pandas as pd
from IPython.display import display # Allows the use of display() for DataFrames
import seaborn as sns
import matplotlib.pyplot as plt
import missingno as msno # module for missing value visualization

# Pretty display for notebooks
%matplotlib inline

<a id="step1"></a>
## step1: import dataset

# import dataset
part1_dataset = pd.read_csv("./Hotel_Reviews.csv")

## step2: have a look at the dataset

# have a look at the dataset 
display(part1_dataset.head(n=2))

# have an overview of all column data type
part1_dataset.info()

<a id="step3"></a>
## step3: convert the column Review_Date to datetime data type.

# As we can tell, the Review_date is an object type, we need to convert it into datetime.
part1_dataset.Review_Date = pd.to_datetime(part1_dataset.Review_Date, format="%m/%d/%Y")

<a id="step4"></a>
## step4: have a look at the descriptive statistical summary

# have a look at the descriptive statistical summary
part1_dataset.describe(include='all')

'''
# if we want to show up all the descriptive summary for both categorical and numerical features.
dataset.describe(include="all")
'''

<a id="step5"></a>
## step5: save the dataset to pickle data for later use.

# save the dataset as part1_dataset
part1_dataset.to_pickle("part1_dataset.pickle")
